webhooks: Add `parseHook` functions
Part of #2. This is in preparation for removing all the other functions
and using only the `parseHook` function in the `webhook.createBuilder`
and `webhook.handleWebhook` functions.

Propagating a single object format from `parseHook` will eliminate the
current coupling between the GitHub format and other parts of the code,
namely `SiteBuilder.launchBuilder` and `Options`.